polish GUINumeric
